refactor(tools): organize tools with tsx and remove zx
