+++ package.json
@@ -7,23 +7,26 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "start": "zx tools/watch.mjs",
-    "build": "zx tools/build.mjs",
+    "start": "tsx tools/watch.ts",
+    "build": "tsx tools/build.ts",
     "lint": "yarn ~textlint",
     "~textlint": "textlint 'aio-ja/**/!(*.en).md'",
     "test": "npm run lint && yarn test:patch",
     "test:patch": "git apply -v --check --directory origin ./tools/git-patch/*.patch",
-    "update-origin": "zx tools/update-origin.mjs"
+    "update-origin": "tsx tools/update-origin.ts"
   },
   "packageManager": "yarn@1.22.10",
   "devDependencies": {
+    "@types/node": "20.14.10",
+    "chalk": "5.3.0",
     "chokidar": "3.6.0",
     "execa": "^9.3.0",
+    "globby": "14.0.2",
     "prh-rules": "prh/rules",
     "textlint": "^14.0.4",
     "textlint-filter-rule-comments": "^1.2.2",
     "textlint-rule-prh": "^6.0.0",
     "tree-kill": "^1.2.2",
-    "zx": "8.1.3"
+    "tsx": "^4.16.2"
   }
 }
