+++ tools/lib/fsutils.ts
@@ -1,33 +1,32 @@
-import { access, cp, mkdir, readFile, rm, writeFile } from 'node:fs/promises';
+import { access, cp, readFile, rm, writeFile } from 'node:fs/promises';
 
-export async function rmrf(path) {
+export async function rmrf(path: string) {
   try {
     await rm(path, { recursive: true });
   } catch {}
 }
 
-export async function exists(path) {
+export async function exists(path: string) {
   return await access(path)
     .then(() => true)
     .catch(() => false);
 }
 
-export async function initDir(path) {
-  await rmrf(path);
-  await mkdir(path, { recursive: true });
-}
-
-export async function cpRf(src, dest) {
+export async function cpRf(src: string, dest: string) {
   await cp(src, dest, { recursive: true, force: true });
 }
 
-export async function sed(path, pattern, replacement) {
+export async function replaceAllInFile(
+  path: string,
+  pattern: string | RegExp,
+  replacement: string
+) {
   const content = await readFile(path, 'utf-8');
   const newContent = content.replaceAll(pattern, replacement);
   await writeFile(path, newContent, 'utf-8');
 }
 
-export async function rename(oldPath, newPath) {
+export async function rename(oldPath: string, newPath: string) {
   await cp(oldPath, newPath);
   await rm(oldPath);
 }
