+++ tools/lib/workspace.ts
@@ -1,0 +1,7 @@
+import { resolve } from 'node:path';
+
+export const rootDir = new URL('../../', import.meta.url).pathname;
+export const adevJaDir = resolve(rootDir, 'adev-ja');
+export const aiojaDir = resolve(rootDir, 'aio-ja');
+export const buildDir = resolve(rootDir, 'build');
+export const buildOutputDir = resolve(buildDir, 'dist/bin/adev/build/browser');
