+++ tools/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "ES2022",
+    "types": ["node"],
+    "allowSyntheticDefaultImports": true
+  },
+  "include": ["lib", "build.ts", "watch.ts", "update-origin.ts"]
+}
